+++ src/main/java/invoker54/reviveme/common/potion/FallenPotionEffect.java
@@ -24,6 +24,7 @@ import java.util.UUID;
 
 public class FallenPotionEffect extends Effect {
     public static List<UUID> warnedPlayers = new ArrayList<>();
+    public static boolean bypassConfig = false;
 
     public static final int effectColor = new Color(35, 5, 5,255).getRGB();
 
@@ -58,7 +59,7 @@ public class FallenPotionEffect extends Effect {
             FallenCapability cap = FallenCapability.get(entity);
             if (cap.isFallen()) return false;
 
-            if (!completed && !ReviveMeConfig.canRemovePenaltyTimer && !((PlayerEntity) entity).isCreative()){
+            if (!bypassConfig && !completed && !ReviveMeConfig.canRemovePenaltyTimer && !((PlayerEntity) entity).isCreative()){
                 if (!warnedPlayers.contains(entity.getUUID())){
                     ((PlayerEntity) entity).displayClientMessage(InvoText.translate("effect.reviveme.fallen_effect.cant_remove").getText(), false);
                     warnedPlayers.add(entity.getUUID());