+++ src/main/java/invoker54/reviveme/mixin/ServerPlayerMixin.java
@@ -4,6 +4,7 @@ import com.mojang.authlib.GameProfile;
 import invoker54.invocore.common.ModLogger;
 import invoker54.reviveme.common.capability.FallenCapability;
 import invoker54.reviveme.common.config.ReviveMeConfig;
+import invoker54.reviveme.common.event.FallEvent;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.entity.player.ServerPlayerEntity;
@@ -69,7 +70,7 @@ public abstract class ServerPlayerMixin extends PlayerEntity {
             },
             cancellable = true)
     private synchronized void hurtStart(DamageSource damageSource, float damage, CallbackInfoReturnable<Boolean> cir) throws InterruptedException {
-        if (damageSource.isBypassInvul()) return;
+        if (FallEvent.canBypassReviveMe(damageSource, false)) return;
         if (revive_Me$getCap() == null) return;
         if (!revive_Me$getCap().isFallen()) return;