+++ changelog.md
@@ -1,3 +1,20 @@
+# 1.16.5-1.17.3
+**Additions**
+
+`-(Config) Damage_Source_Whitelist - What damage sources that can bypass fallen state`
+
+`-(Config) Can_Pick_Up_items - if fallen players can grab items`
+
+**Changes**
+
+`-FixCommand and ReviveCommand now accounts for multiple players`
+
+`-You can now remove Penalty timer with effect command regardless of Can_Remove_Penalty_Timer config`
+
+**Fixes**
+
+`-Revive effects can now be properly hidden`
+
 # 1.16.5-1.17.1
 
 <div><pre><strong>Additions</strong><br>-(Config) Max_Total_Revives - Max amount of time you can be revived<br>-(Config) Max_Player_Revives - Max amount of times someone else can revive you<br>-(Config) Revive_Radius - Radius teammates must be within for you to enter the fallen state<br><br><strong>Changes</strong><br>-(Config) Time_Left can now go to -1 to disable it<br>-(Config) Max_Self_Revives has been moved to General Fallen State Settings<br>-Fallen Effect now warns people if it can be removed if they leave and rejoin a world<br>-Removed Cancel_Revive_On_Damage since it wasn't being used<br>-Every effect can now be hidden through config<br>-Blocked keybinds are now checked once you enter the fallen state<br><br><strong>Fixes</strong><br>-Revive Me! now removes effects one at a time instead of all at once<br>-Fallen Penalty now happens on the 2nd down<br>-OpenGL Key error no longer spams in log<br>-Reviver's capability is properly synced on successful revive</pre></div>